refactor(test): split job-result tests if needed
